Modify zone-list.sh to read input and output dirs as parameters
Add checks that print out error messages if input parameters
are bad. Print usage if script gets option -h, or if paramters
are bad.

+++ scripts/zone-list.sh
@@ -1,27 +1,75 @@
 #!/bin/bash -e
 
-# set -o xtrace
+function usage() {
+    echo "Usage ${0##*/} [TZDATA DIR] [BUILD DIR]
+Builds the tz database from sources in [TZDATA DIR]
+without links to directory [BUILD DIR]/zones-without-links
+and prints out all time zones in the resulting database.
+See man zic for more information about time zone links.
+This script expects that the directory [BUILD DIR] contains
+the files
 
-# This script has to be executed in the build directory
+  iso8601    Custom ISO 8016 time zones in the same format
+             as the time zone files in [TZDATA DIR]
+  yearistype Executable to check year type, see man zic
 
-zic=bin/zic
-yearistype=bin/yearistype
-output=zones-listing
+Options:
+  -h         Print this message"
+}
 
-test -x $zic
-test -x $yearistype
+if [ "$1" == "-h" ]; then
+    usage
+    exit
+fi
+
+if [ $# != 2 ]; then
+    echo "${0##*/}: wrong number of arguments"
+    usage
+    exit 1
+fi
+
+if [ ! -d $1 ]; then
+    echo "${0##*/}: the directory $1 does not exist"
+    usage
+    exit 1
+fi
+
+if [ ! -d $2 ]; then
+    echo "${0##*/}: the directory $2 does not exist"
+    usage
+    exit 1
+fi
+
+src_dir=$1
+build_dir=$2
+
+zic=/usr/sbin/zic
+yearistype=$build_dir/yearistype
+output=$build_dir/zones-without-links
+
+if [ ! -x $zic ]; then
+    echo "${0##*/}: $zic does not exist or is not executable"
+    exit 1
+fi
+
+if [ ! -x $yearistype ]; then
+    echo "${0##*/}: $yearistype does not exist or is not executable"
+    exit 1
+fi
+
 rm -rf $output && mkdir -p $output
 
 zic_cmd="$zic -L /dev/null -y $yearistype "
 
-input="africa antarctica asia australasia europe northamerica southamerica iso8601"
+input="$src_dir/africa $src_dir/antarctica $src_dir/asia $src_dir/australasia $src_dir/europe $src_dir/northamerica $src_dir/southamerica"
+input="$input $build_dir/iso8601"
 
 for continent in $input ; do
   test -f $continent
 done
 
-cat $input | pcregrep -v '^\s*Link\s+' | tee aaaa | $zic_cmd -d $output 2> stderr -
-cat stderr | grep -v "time zone abbreviation differs from POSIX standard" >&2 || true
+cat $input | pcregrep -v '^\s*Link\s+' | tee $build_dir/aaaa | $zic_cmd -d $output 2> $build_dir/stderr -
+cat $build_dir/stderr | grep -v "time zone abbreviation differs from POSIX standard" >&2 || true
 
 {
   for f in $(cd $output ; echo *) ; do
@@ -31,7 +79,4 @@ cat stderr | grep -v "time zone abbreviation differs from POSIX standard" >&2 ||
       echo $f
     fi
   done
-
-#  ( cat zone.tab | pcregrep -v '^\s*#' | awk '{print $3}' )
-
 } | sort | uniq
